backend_update/triviaback/requirements.txt
asgiref
certifi
charset-normalizer
colorama
decorator
Django
django-allauth
django-cors-headers
djangorestframework
djangorestframework-simplejwt
idna
imageio
imageio-ffmpeg
moviepy
numpy
pillow
proglog
PyJWT
python-dotenv
requests
setuptools
six
sqlparse
tqdm
tzdata
